refactor: context분리
